Extension for printing the source line responsible for a failed assertion
This greatly speeds up disecting just what went wrong if a test fails. I
believe this is in line with should.js' goals of being expressive and helpful,
but would love to hear your feedback.

This extension isn't enabled in node or browser as this will break tests. If
this is thought to be general enough to be enabled by default, the tests can
be fixed to match the new output.